refactor: move OpenAPIContext to datastructures.py
